resume_inspector/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
spacy==3.7.2
python-dateutil==2.8.2
requests==2.31.0
Werkzeug==2.3.7
python-docx==0.8.11
PyPDF2==3.0.1
nltk
numpy
pandas
python-dotenv
gunicorn 